extract useEffect to hook
